Add lid-close sleep, configurable sleep buttons, and new config options
Inject a configurable GBA button combo on lid close to trigger sleep in
games that support it. Mute audio, stop frame capture, and dim the power
LED while the lid is closed. Add colorOverride and sleepButtons config
options. Change defaultSave from u16 to u8.

+++ include/arm11/save_type.h
@@ -38,7 +38,7 @@ static_assert(sizeof(GbaDbEntry) == 28, "Error: GBA DB entry struct is not packe
 
 
 
-u16 detectSaveType(const u32 romSize, const u16 defaultSave);
+u16 detectSaveType(const u32 romSize, const u8 defaultSave);
 u16 getSaveType(const OafConfig *const cfg, const u32 romSize, const char *const savePath);
 
 #ifdef __cplusplus
